+++ src/components/cards/product/index.tsx
@@ -37,8 +37,8 @@ const ProductCard = (props: any) => {
           <Link href={`/catalog/${product.id}`}>
             {" "}
             <div className={s.card_title}>{product?.title.slice(0, 26)}</div>
-            <div className={s.card_model}>{product?.title.slice(28)}</div>
             <div className={s.card__producer}>_{product?.producer}</div>
+            <div className={s.card_model}>{product?.title.slice(28)}</div>
           </Link>
           <div className={s.price_cart_container}>
             <div>
