+++ src/pages/cart/index.module.scss
@@ -4,6 +4,7 @@
     width: 100%;
     @include flex-column-center;
 }
+
 .empty_cart_wrapper {
     width: 100%;
     display: flex;
@@ -26,7 +27,7 @@
     flex-direction: column;
     align-items: center;
     justify-content: center;
-    
+
 }
 
 .items_wrapper {
@@ -83,6 +84,7 @@
     width: 100%;
     display: flex;
     justify-content: space-between;
+    margin-bottom: 20px;
 }
 
 .bottom_wrapper {
@@ -105,6 +107,7 @@
         width: 100%;
         display: flex;
         justify-content: space-between;
+        flex-wrap: wrap;
         font-weight: 600;
         font-size: 24px;
         line-height: 26px;
@@ -130,14 +133,20 @@
     .bottom_btn_container {
         width: 100%;
         display: flex;
+        gap: 20px;
+        flex-wrap: wrap;
         justify-content: space-between;
-        align-items: center;
+        align-items: end;
+
+        @media screen and (max-width: 923px) {
+            justify-content: center;
+        }
 
         .payment_link_down {
             text-align: center;
-            margin-right: 20px;
+            // margin-right: 20px;
             text-decoration: none;
-            padding: 16px 45px;
+            padding: 16px 40px;
             color: white;
             margin-top: 20px;
             background: #C90000;
@@ -154,17 +163,23 @@
 
 .contiue_shopping_btn {
     text-align: center;
-    margin-right: 20px;
+    // margin-right: 20px;
     text-decoration: none;
-    padding: 16px 45px;
-    color:$primary-color;
+    padding: 16px 32px;
+    color: $primary-color;
 
     margin-top: 20px;
     border: 2px solid #C90000;
 
+    @media screen and (max-width: 923px) {
+        margin-top: 0;
+    }
+
+
+
     &:hover {
-        
-        color:$primary-bg-color;
+
+        color: $primary-bg-color;
         background: $primary-color;
     }
 }
@@ -175,29 +190,38 @@
     flex-direction: column;
     align-items: flex-start;
     margin-top: 20px;
-}
 
-.customHeader {
-    width: 100%;
-    padding: 0;
-    border: none;
-    background: none;
-}
+    .accordion_delivery {
+        display: flex;
+       width: 100%; 
+       padding: 0 5%;
+    }
 
-.customBody {
-    align-self: flex-start;
-    background:$primary-color;
-}
-.form {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    input {
+    .customHeader {
+        width: 100%;
+        padding: 0;
         border: none;
-        border-bottom: 1px solid #C90000;
-        padding: 10px 5px;
-        &:focus {
-            border:none!important;
+        background: none;
+    }
+
+    .customBody {
+        align-self: flex-start;
+        background: $primary-color;
+    }
+
+    .form {
+        width: 100%;
+        display: flex;
+        flex-direction: column;
+
+        input {
+            border: none;
+            border-bottom: 1px solid #C90000;
+            padding: 10px 5px;
+
+            &:focus {
+                border: none !important;
+            }
         }
     }
-} 
+}
